requirements.txt
openai
python-dotenv
PyAudio
pydub
wave